Move PrometheusExporter to contrib package
